Extract byte order agnostic BaseWritableBufferImpl
